+++ Sandbox/GLSLBook/GLSLBookPlugins.cpp
@@ -8,11 +8,12 @@
 #include "GLSLBook/PostEffectPipeline.h"
 #include "GLSLBook/ImageProcessing/EdgePipeline.h"
 #include "GLSLBook/ImageProcessing/BlurPipeline.h"
+#include "GLSLBook/ImageProcessing/BloomPipeline.h"
 
 using namespace Zelo::Core::RHI;
 
 const std::string &EdgePipelinePlugin::getName() const {
-    static std::string s = "PostEffect";
+    static std::string s = "EdgePipelinePlugin";
     return s;
 }
 
@@ -35,7 +36,7 @@ void EdgePipelinePlugin::uninstall() {
 }
 
 const std::string &BlurPipelinePlugin::getName() const {
-    static std::string s = "PostEffect";
+    static std::string s = "BlurPipelinePlugin";
     return s;
 }
 
@@ -54,5 +55,28 @@ void BlurPipelinePlugin::shutdown() {
 }
 
 void BlurPipelinePlugin::uninstall() {
+    RenderSystem::getSingletonPtr()->resetRenderPipeline();
+}
+
+const std::string &BloomPipelinePlugin::getName() const {
+    static std::string s = "BloomPipelinePlugin";
+    return s;
+}
+
+void BloomPipelinePlugin::install() {
+    auto renderPipeline = std::make_unique<BloomPipeline>();
+    renderPipeline->initialize();
+    RenderSystem::getSingletonPtr()->setRenderPipeline(std::move(renderPipeline));
+}
+
+void BloomPipelinePlugin::initialise() {
+    // do nothing
+}
+
+void BloomPipelinePlugin::shutdown() {
+    // do nothing
+}
+
+void BloomPipelinePlugin::uninstall() {
     RenderSystem::getSingletonPtr()->resetRenderPipeline();
 }
