+++ Sandbox/GLSLBook/GLSLBookPlugins.h
@@ -31,3 +31,16 @@ public:
 
     void uninstall() override;
 };
+
+class BloomPipelinePlugin : public Plugin {
+public:
+    const std::string &getName() const override;;
+
+    void install() override;
+
+    void initialise() override;
+
+    void shutdown() override;
+
+    void uninstall() override;
+};
