+++ Sandbox/GLSLBook/ImageProcessing/BloomPipeline.h
@@ -1,9 +1,7 @@
-// Bloom.h
+// BloomPipeline.h
 // created on 2021/3/29
 // author @zoloypzuo
-
-#ifndef ZELOENGINE_FORWARDRENDERER_H
-#define ZELOENGINE_FORWARDRENDERER_H
+#pragma once
 
 #include "ZeloPrerequisites.h"
 #include "ZeloGLPrerequisites.h"
@@ -12,28 +10,23 @@
 #include "Renderer/OpenGL/Resource/GLSLShaderProgram.h"
 #include "Renderer/OpenGL/Drawable/Line.h"
 #include "Renderer/OpenGL/Buffer/GLFramebuffer.h"
+#include "Renderer/OpenGL/Pipeline/ForwardPipeline.h"
+#include "GLSLBook/Drawable/Quad.h"
 
-class Bloom : public  Zelo::Core::RHI::RenderPipeline {
+
+class BloomPipeline : public Zelo::Renderer::OpenGL::ForwardPipeline{
 public:
-    Bloom();
+    BloomPipeline();
 
-    ~Bloom() override;
+    ~BloomPipeline() override;
 
     void render(const Zelo::Core::ECS::Entity &scene) const override;
 
     void initialize() override;
 
 protected:
-    void createShaders();
-
-    std::unique_ptr<GLSLShaderProgram> m_forwardAmbient;
-    std::unique_ptr<GLSLShaderProgram> m_forwardDirectional;
-    std::unique_ptr<GLSLShaderProgram> m_forwardPoint;
-    std::unique_ptr<GLSLShaderProgram> m_forwardSpot;
-
     // post effect
     std::unique_ptr<GLSLShaderProgram> m_postShader;
     std::unique_ptr<Zelo::GLFramebuffer> m_renderFbo, m_fbo1, m_fbo2;
+    Quad m_quad{};
 };
-
-#endif //ZELOENGINE_FORWARDRENDERER_H
